requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
confluent-kafka==2.8.2
idna==3.10
py4j==0.10.9.7
pyspark==3.5.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.3.0
